Make Pid type part of public API
This change moves the maps::Pid type into lib.rs and makes it part of
the public API surface. Upcoming changes will convert existing APIs to
use it instead of "raw" u32 values or Options of them.
