+++ src/lib.rs
@@ -41,6 +41,10 @@ mod util;
 #[allow(unused)]
 mod zip;
 
+use std::fmt::Display;
+use std::fmt::Formatter;
+use std::fmt::Result as FmtResult;
+use std::num::NonZeroU32;
 use std::path::PathBuf;
 
 #[cfg(feature = "symbolize")]
@@ -64,6 +68,29 @@ pub use symbolize::SymbolizerFeature;
 use c_api::*;
 
 pub type Addr = usize;
+
+
+/// An enumeration identifying a process.
+#[derive(Clone, Copy, Debug)]
+pub enum Pid {
+    Slf,
+    Pid(NonZeroU32),
+}
+
+impl Display for Pid {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        match self {
+            Self::Slf => write!(f, "self"),
+            Self::Pid(pid) => write!(f, "{pid}"),
+        }
+    }
+}
+
+impl From<u32> for Pid {
+    fn from(pid: u32) -> Self {
+        NonZeroU32::new(pid).map(Pid::Pid).unwrap_or(Pid::Slf)
+    }
+}
 
 
 /// The context of an address finding request.
